prevent users not running TIA and then trying to use prioritized tests
