+++ agent/src/test/java/com/teamscale/jacoco/agent/testimpact/TestwiseCoverageAgentTest.java
@@ -16,6 +16,7 @@ import com.teamscale.report.testwise.model.builder.FileCoverageBuilder;
 import com.teamscale.report.testwise.model.builder.TestCoverageBuilder;
 import com.teamscale.tia.client.RunningTest;
 import com.teamscale.tia.client.TestRun;
+import com.teamscale.tia.client.TestRunWithSuggestions;
 import com.teamscale.tia.client.TiaAgent;
 import okhttp3.HttpUrl;
 import org.junit.jupiter.api.Test;
@@ -80,7 +81,7 @@ public class TestwiseCoverageAgentTest {
 
 		TiaAgent agent = new TiaAgent(false, HttpUrl.get("http://localhost:" + port));
 
-		TestRun testRun = agent.startTestRun(availableTests);
+		TestRunWithSuggestions testRun = agent.startTestRun(availableTests);
 		assertThat(testRun.getPrioritizedTests()).hasSize(1);
 		assertThat(testRun.getPrioritizedTests().get(0).tests).hasSize(1);
 		PrioritizableTest test = testRun.getPrioritizedTests().get(0).tests.get(0);
